helmrepo and appsub content removal (#930)
* helmrepo and appsub content removal


* additional helmrepo and appsub content removal


* prune resources on upgrade


* Remove dependencies


* Cleanup helmrepo and channel resources after upgrade


* Finalize volsync and address nil MCE


* Add uninstall.go test coverage
