+++ controllers/common.go
@@ -19,11 +19,9 @@ import (
 	olmv1 "github.com/operator-framework/api/pkg/operators/v1"
 
 	configv1 "github.com/openshift/api/config/v1"
-	"github.com/openshift/library-go/pkg/operator/resource/resourcemerge"
 	subv1alpha1 "github.com/operator-framework/api/pkg/operators/v1alpha1"
 	searchv2v1alpha1 "github.com/stolostron/search-v2-operator/api/v1alpha1"
 	apixv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	subhelmv1 "open-cluster-management.io/multicloud-operators-subscription/pkg/apis/apps/helmrelease/v1"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -34,14 +32,10 @@ import (
 	utils "github.com/stolostron/multiclusterhub-operator/pkg/utils"
 
 	operatorv1 "github.com/stolostron/multiclusterhub-operator/api/v1"
-	"github.com/stolostron/multiclusterhub-operator/pkg/channel"
-	"github.com/stolostron/multiclusterhub-operator/pkg/helmrepo"
 	"github.com/stolostron/multiclusterhub-operator/pkg/manifest"
 	"github.com/stolostron/multiclusterhub-operator/pkg/multiclusterengine"
-	"github.com/stolostron/multiclusterhub-operator/pkg/subscription"
 	"github.com/stolostron/multiclusterhub-operator/pkg/version"
 
-	"k8s.io/apimachinery/pkg/api/equality"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/types"
@@ -93,201 +87,6 @@ func (r *MultiClusterHubReconciler) ensureDeployment(m *operatorv1.MultiClusterH
 		return ctrl.Result{}, err
 	}
 
-	// Validate object based on name
-	var desired *appsv1.Deployment
-	var needsUpdate bool
-
-	switch found.Name {
-	case helmrepo.HelmRepoName:
-		desired, needsUpdate = helmrepo.ValidateDeployment(m, r.CacheSpec.ImageOverrides, dep, found)
-	default:
-		r.Log.Info("Could not validate deployment; unknown name")
-		return ctrl.Result{}, nil
-	}
-
-	if needsUpdate {
-		err = r.Client.Update(context.TODO(), desired)
-		if err != nil {
-			r.Log.Error(err, "Failed to update Deployment.")
-			return ctrl.Result{}, err
-		}
-		// Spec updated - return
-		return ctrl.Result{}, nil
-	}
-	return ctrl.Result{}, nil
-}
-
-func (r *MultiClusterHubReconciler) ensureService(m *operatorv1.MultiClusterHub, s *corev1.Service) (ctrl.Result, error) {
-	svlog := r.Log.WithValues("Service.Namespace", s.Namespace, "Service.Name", s.Name)
-
-	found := &corev1.Service{}
-	err := r.Client.Get(context.TODO(), types.NamespacedName{
-		Name:      s.Name,
-		Namespace: m.Namespace,
-	}, found)
-	if err != nil && errors.IsNotFound(err) {
-
-		// Create the service
-		err = r.Client.Create(context.TODO(), s)
-
-		if err != nil {
-			// Creation failed
-			svlog.Error(err, "Failed to create new Service")
-			return ctrl.Result{}, err
-		}
-
-		// Creation was successful
-		svlog.Info("Created a new Service")
-		condition := NewHubCondition(operatorv1.Progressing, metav1.ConditionTrue, NewComponentReason, "Created new resource")
-		SetHubCondition(&m.Status, *condition)
-		return ctrl.Result{}, nil
-
-	} else if err != nil {
-		// Error that isn't due to the service not existing
-		svlog.Error(err, "Failed to get Service")
-		return ctrl.Result{}, err
-	}
-
-	modified := resourcemerge.BoolPtr(false)
-	existingCopy := found.DeepCopy()
-	resourcemerge.EnsureObjectMeta(modified, &existingCopy.ObjectMeta, s.ObjectMeta)
-	selectorSame := equality.Semantic.DeepEqual(existingCopy.Spec.Selector, s.Spec.Selector)
-
-	typeSame := false
-	requiredIsEmpty := len(s.Spec.Type) == 0
-	existingCopyIsCluster := existingCopy.Spec.Type == corev1.ServiceTypeClusterIP
-	if (requiredIsEmpty && existingCopyIsCluster) || equality.Semantic.DeepEqual(existingCopy.Spec.Type, s.Spec.Type) {
-		typeSame = true
-	}
-
-	if selectorSame && typeSame && !*modified {
-		return ctrl.Result{}, nil
-	}
-
-	existingCopy.Spec.Selector = s.Spec.Selector
-	existingCopy.Spec.Type = s.Spec.Type
-	err = r.Client.Update(context.TODO(), existingCopy)
-	if err != nil {
-		svlog.Error(err, "Failed to update Service")
-		return ctrl.Result{}, err
-	}
-	return ctrl.Result{}, nil
-}
-
-func (r *MultiClusterHubReconciler) ensureChannel(m *operatorv1.MultiClusterHub, u *unstructured.Unstructured) (ctrl.Result, error) {
-	selog := r.Log.WithValues("Channel.Namespace", u.GetNamespace(), "Channel.Name", u.GetName())
-
-	found := &unstructured.Unstructured{}
-	found.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "apps.open-cluster-management.io",
-		Kind:    "Channel",
-		Version: "v1",
-	})
-	err := r.Client.Get(context.TODO(), types.NamespacedName{
-		Name:      u.GetName(),
-		Namespace: m.Namespace,
-	}, found)
-	if err != nil && errors.IsNotFound(err) {
-		// Create the Channel
-		err = r.Client.Create(context.TODO(), u)
-		if err != nil {
-			// Creation failed
-			selog.Error(err, "Failed to create new Channel")
-			return ctrl.Result{}, err
-		}
-
-		// Creation was successful
-		selog.Info("Created a new Channel")
-		condition := NewHubCondition(operatorv1.Progressing, metav1.ConditionTrue, NewComponentReason, "Created new resource")
-		SetHubCondition(&m.Status, *condition)
-		return ctrl.Result{}, nil
-
-	} else if err != nil {
-		// Error that isn't due to the Channel not existing
-		selog.Error(err, "Failed to get Channel")
-		return ctrl.Result{}, err
-	}
-
-	updated, needsUpdate := channel.Validate(m, found)
-	if needsUpdate {
-		selog.Info("Updating channel")
-		err = r.Client.Update(context.TODO(), updated)
-		if err != nil {
-			// Update failed
-			selog.Error(err, "Failed to update channel")
-			return ctrl.Result{}, err
-		}
-		return ctrl.Result{}, nil
-	}
-
-	return ctrl.Result{}, nil
-}
-
-func (r *MultiClusterHubReconciler) ensureSubscription(m *operatorv1.MultiClusterHub, u *unstructured.Unstructured) (ctrl.Result, error) {
-	obLog := r.Log.WithValues("Namespace", u.GetNamespace(), "Name", u.GetName(), "Kind", u.GetKind())
-
-	if utils.ProxyEnvVarsAreSet() {
-		u = addProxyEnvVarsToSub(u)
-	}
-
-	found := &unstructured.Unstructured{}
-	found.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "apps.open-cluster-management.io",
-		Kind:    "Subscription",
-		Version: "v1",
-	})
-	// Try to get API group instance
-	err := r.Client.Get(context.TODO(), types.NamespacedName{
-		Name:      u.GetName(),
-		Namespace: u.GetNamespace(),
-	}, found)
-	if err != nil && errors.IsNotFound(err) {
-
-		err := r.Client.Create(context.TODO(), u)
-		if err != nil {
-			// Creation failed
-			obLog.Error(err, "Failed to create new instance")
-			return ctrl.Result{}, err
-		}
-
-		// Creation was successful
-		obLog.Info("Created new object")
-		condition := NewHubCondition(operatorv1.Progressing, metav1.ConditionTrue, NewComponentReason, "Created new resource")
-		SetHubCondition(&m.Status, *condition)
-		return ctrl.Result{}, nil
-
-	} else if err != nil {
-		// Error that isn't due to the resource not existing
-		obLog.Error(err, "Failed to get subscription")
-		return ctrl.Result{}, err
-	}
-
-	// Validate object based on type
-	updated, needsUpdate := subscription.Validate(found, u)
-	if needsUpdate {
-		obLog.Info("Updating subscription")
-		// Update the resource. Skip on unit test
-		err = r.Client.Update(context.TODO(), updated)
-		if err != nil {
-			// Update failed
-			obLog.Error(err, "Failed to update object")
-			return ctrl.Result{}, err
-		}
-
-		// Spec updated - return
-		return ctrl.Result{}, nil
-	}
-
-	return ctrl.Result{}, nil
-}
-
-func (r *MultiClusterHubReconciler) ensureNoSubscription(m *operatorv1.MultiClusterHub, u *unstructured.Unstructured) (ctrl.Result, error) {
-	subLog := r.Log.WithValues("Namespace", u.GetNamespace(), "Name", u.GetName(), "Kind", u.GetKind())
-	_, err := r.uninstall(m, u)
-	if err != nil {
-		subLog.Error(err, "Failed to uninstall subscription")
-		return ctrl.Result{}, err
-	}
 	return ctrl.Result{}, nil
 }
 
@@ -305,25 +104,6 @@ func (r *MultiClusterHubReconciler) ensureNoDeployment(m *operatorv1.MultiCluste
 	_, err = r.uninstall(m, u)
 	if err != nil {
 		dplog.Error(err, "Failed to uninstall subscription")
-		return ctrl.Result{}, err
-	}
-	return ctrl.Result{}, nil
-}
-
-func (r *MultiClusterHubReconciler) ensureNoService(m *operatorv1.MultiClusterHub, s *corev1.Service) (ctrl.Result, error) {
-	svlog := r.Log.WithValues("Service.Namespace", s.Namespace, "Service.Name", s.Name)
-
-	unstructuredMap, err := runtime.DefaultUnstructuredConverter.ToUnstructured(s)
-	if err != nil {
-		r.Log.Error(err, "Failed to unmarshal deployment")
-		return ctrl.Result{}, err
-	}
-	u := &unstructured.Unstructured{Object: unstructuredMap}
-	u.SetGroupVersionKind(schema.GroupVersionKind{Group: "", Kind: "Service", Version: "v1"})
-
-	_, err = r.uninstall(m, u)
-	if err != nil {
-		svlog.Error(err, "Failed to uninstall subscription")
 		return ctrl.Result{}, err
 	}
 	return ctrl.Result{}, nil
@@ -686,25 +466,6 @@ func (r *MultiClusterHubReconciler) listDeployments(namespaces []string) ([]*app
 	return ret, nil
 }
 
-// listHelmReleases gets all helmreleases in the given namespaces
-func (r *MultiClusterHubReconciler) listHelmReleases(namespaces []string) ([]*subhelmv1.HelmRelease, error) {
-	var ret []*subhelmv1.HelmRelease
-
-	for _, n := range namespaces {
-		hrList := &subhelmv1.HelmReleaseList{}
-		err := r.Client.List(context.TODO(), hrList, client.InNamespace(n))
-		if err != nil && !errors.IsNotFound(err) {
-			return nil, err
-		}
-
-		for i := 0; i < len(hrList.Items); i++ {
-			ret = append(ret, &hrList.Items[i])
-		}
-	}
-
-	return ret, nil
-}
-
 // listCustomResources gets custom resources the installer observes
 func (r *MultiClusterHubReconciler) listCustomResources(m *operatorv1.MultiClusterHub) (map[string]*unstructured.Unstructured, error) {
 	ret := make(map[string]*unstructured.Unstructured)
@@ -749,18 +510,6 @@ func (r *MultiClusterHubReconciler) listCustomResources(m *operatorv1.MultiClust
 	return ret, nil
 }
 
-// filterDeploymentsByRelease returns a subset of deployments with the release label value
-func filterDeploymentsByRelease(deploys []*appsv1.Deployment, releaseLabel string) []*appsv1.Deployment {
-	var labeledDeps []*appsv1.Deployment
-	for i := range deploys {
-		anno := deploys[i].GetAnnotations()
-		if anno["meta.helm.sh/release-name"] == releaseLabel {
-			labeledDeps = append(labeledDeps, deploys[i])
-		}
-	}
-	return labeledDeps
-}
-
 // addInstallerLabel adds the installer name and namespace to a deployment's labels
 // so it can be watched. Returns false if the labels are already present.
 func addInstallerLabel(d *appsv1.Deployment, name string, ns string) bool {
@@ -797,35 +546,6 @@ func addInstallerLabelSecret(d *corev1.Secret, name string, ns string) bool {
 	return updated
 }
 
-// getAppSubOwnedHelmReleases gets a subset of helmreleases created by the appsubs
-func getAppSubOwnedHelmReleases(allHRs []*subhelmv1.HelmRelease, appsubs []types.NamespacedName) []*subhelmv1.HelmRelease {
-	subMap := make(map[string]bool)
-	for _, s := range appsubs {
-		subMap[s.Name] = true
-	}
-
-	var ownedHRs []*subhelmv1.HelmRelease
-	for _, hr := range allHRs {
-		// check if this is one of our helmreleases
-		owner := hr.OwnerReferences[0].Name
-		if subMap[owner] {
-			ownedHRs = append(ownedHRs, hr)
-
-		}
-	}
-	return ownedHRs
-}
-
-// getHelmReleaseOwnedDeployments gets a subset of deployments created by the helmreleases
-func getHelmReleaseOwnedDeployments(allDeps []*appsv1.Deployment, hrList []*subhelmv1.HelmRelease) []*appsv1.Deployment {
-	var mchDeps []*appsv1.Deployment
-	for _, hr := range hrList {
-		hrDeployments := filterDeploymentsByRelease(allDeps, hr.Name)
-		mchDeps = append(mchDeps, hrDeployments...)
-	}
-	return mchDeps
-}
-
 // labelDeployments updates deployments with installer labels if not already present
 func (r *MultiClusterHubReconciler) labelDeployments(hub *operatorv1.MultiClusterHub, dList []*appsv1.Deployment) error {
 	for _, d := range dList {
@@ -840,50 +560,6 @@ func (r *MultiClusterHubReconciler) labelDeployments(hub *operatorv1.MultiCluste
 		}
 	}
 	return nil
-}
-
-// ensureSubscriptionOperatorIsRunning verifies that the subscription operator that manages helm subscriptions exists and
-// is running. This validation is only intended to run during upgrade and when run as an OLM managed deployment
-func (r *MultiClusterHubReconciler) ensureSubscriptionOperatorIsRunning(mch *operatorv1.MultiClusterHub, allDeps []*appsv1.Deployment) (ctrl.Result, error) {
-	// skip check if not upgrading
-	if mch.Status.CurrentVersion == version.Version {
-		return ctrl.Result{}, nil
-	}
-
-	selfDeployment, exists := getDeploymentByName(allDeps, utils.MCHOperatorName)
-	if !exists {
-		// Deployment doesn't exist so this is either being run locally or with unit tests
-		return ctrl.Result{}, nil
-	}
-
-	// skip check if not deployed by OLM
-	if !isACMManaged(selfDeployment) {
-		return ctrl.Result{}, nil
-	}
-
-	subscriptionDeploy, exists := getDeploymentByName(allDeps, utils.SubscriptionOperatorName)
-	if !exists {
-		err := fmt.Errorf("Standalone subscription deployment not found")
-		return ctrl.Result{}, err
-	}
-
-	// Check that the owning CSV version of the deployments match
-	if selfDeployment.GetLabels() == nil || subscriptionDeploy.GetLabels() == nil {
-		r.Log.Info("Missing labels on either MCH operator or subscription operator deployment")
-		return ctrl.Result{RequeueAfter: time.Second * 10}, nil
-	}
-	if selfDeployment.Labels["olm.owner"] != subscriptionDeploy.Labels["olm.owner"] {
-		r.Log.Info("OLM owner labels do not match. Requeuing.", "MCH operator label", selfDeployment.Labels["olm.owner"], "Subscription operator label", subscriptionDeploy.Labels["olm.owner"])
-		return ctrl.Result{RequeueAfter: time.Second * 10}, nil
-	}
-
-	// Check that the standalone subscription deployment is available
-	if successfulDeploy(subscriptionDeploy) {
-		return ctrl.Result{}, nil
-	} else {
-		r.Log.Info("Standalone subscription deployment is not running. Requeuing.")
-		return ctrl.Result{RequeueAfter: time.Second * 10}, nil
-	}
 }
 
 // GetSubscriptionIPInformation retrieves the current subscriptions installplan information for status
@@ -989,6 +665,10 @@ func (r *MultiClusterHubReconciler) waitForMCEReady(ctx context.Context) (ctrl.R
 	existingMCE, err := multiclusterengine.GetManagedMCE(ctx, r.Client)
 	if err != nil {
 		return ctrl.Result{Requeue: true}, err
+	}
+	if existingMCE == nil {
+		r.Log.Info(fmt.Sprintf("Multiclusterengine: %s is not yet present", existingMCE.GetName()))
+		return ctrl.Result{Requeue: true}, nil
 	}
 	if utils.IsUnitTest() {
 		return ctrl.Result{}, nil
