+++ main.go
@@ -53,9 +53,6 @@ import (
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	apiregistrationv1 "k8s.io/kube-aggregator/pkg/apis/apiregistration/v1"
-	policy "open-cluster-management.io/governance-policy-propagator/api/v1"
-	channel "open-cluster-management.io/multicloud-operators-channel/pkg/apis"
-	appsubv1 "open-cluster-management.io/multicloud-operators-subscription/pkg/apis"
 
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -80,12 +77,6 @@ func init() {
 	utilruntime.Must(operatorv1.AddToScheme(scheme))
 
 	utilruntime.Must(searchv2v1alpha1.AddToScheme(scheme))
-
-	utilruntime.Must(policy.AddToScheme(scheme))
-
-	utilruntime.Must(appsubv1.AddToScheme(scheme))
-
-	utilruntime.Must(channel.AddToScheme(scheme))
 
 	utilruntime.Must(apiregistrationv1.AddToScheme(scheme))
